refactor(api-headless-cms): extract request handler so it can be used as standalone method
