+++ packages/api-headless-cms/src/graphql/checkEndpointAccess.ts
@@ -1,0 +1,13 @@
+import { CmsContext } from "~/types";
+import { NotAuthorizedError } from "@webiny/api-security";
+
+export const checkEndpointAccess = async (context: CmsContext): Promise<void> => {
+    const permission = await context.security.getPermission(`cms.endpoint.${context.cms.type}`);
+    if (!permission) {
+        throw new NotAuthorizedError({
+            data: {
+                reason: `Not allowed to access "${context.cms.type}" endpoint.`
+            }
+        });
+    }
+};
