+++ packages/api-headless-cms/src/graphql/formatErrorPayload.ts
@@ -1,0 +1,19 @@
+import WebinyError from "@webiny/error";
+
+export const formatErrorPayload = (error: Error): string => {
+    if (error instanceof WebinyError) {
+        return JSON.stringify({
+            type: "CmsGraphQLWebinyError",
+            message: error.message,
+            code: error.code,
+            data: error.data
+        });
+    }
+
+    return JSON.stringify({
+        type: "Error",
+        name: error.name,
+        message: error.message,
+        stack: error.stack
+    });
+};
