+++ packages/handler-graphql/src/index.ts
@@ -4,5 +4,6 @@ import createGraphQLHandler from "./createGraphQLHandler";
 export * from "./errors";
 export * from "./responses";
 export * from "./plugins";
+export { default as processRequestBody } from "./processRequestBody";
 
 export default (options: HandlerGraphQLOptions = {}) => [createGraphQLHandler(options)];
